fix: type of fk collumns

+++ database/migrations/2024_10_03_124532_create_users_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent();
 
-            $table->unsignedBigInteger('role_id'); 
+            $table->integer('role_id'); 
             $table->foreign('role_id')->references('id')->on('users_roles')->onDelete('cascade');
         });
     }
